hw: extract SPI helpers from hardware.cpp to spi.cpp
